move matchMedia and breakpoint and localstorage to `misc/`
