+++ packages/web/src/event/index.ts
@@ -2,4 +2,4 @@ export * from "./event";
 export * from "./onMouseMove";
 export * from "./onResize";
 export * from "./onScroll";
-export * from './breakpoint'
+export * from '../misc/breakpoint'
