+++ packages/web/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./event";
 export * from "./web";
+export * from "./misc";
