+++ packages/web/src/misc/index.ts
@@ -1,0 +1,3 @@
+
+export * from '../misc/localStorage';
+export * from '../misc/matchMedia'
