+++ packages/web/src/web/index.ts
@@ -1,5 +1,3 @@
 export * from './fetch';
 export * from './webSocket';
-export * from './localStorage';
 export * from './intersectionObserver'
-export * from './matchMedia'
